+++ suggest/CHANGELOG.md
@@ -17,6 +17,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [0.0.2] - 2021-07-15
+
+### Added
+
+* Added `suggestion` string property on `StepDocument`
+  ([#1657](https://github.com/cucumber/common/pull/1657)
+   [aslakhellesoy])
+
+### Changed
+
+* The `StepDocument` type has changed to `{ suggestion: string, segments: StepSegments }`
+  ([#1657](https://github.com/cucumber/common/pull/1657)
+   [aslakhellesoy])
+
 ## [0.0.1]
 
 ### Added
@@ -24,7 +38,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * First release!
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/common/compare/suggest/v0.0.1...main
+[Unreleased]: https://github.com/cucumber/common/compare/suggest/v0.0.2...main
+[0.0.2]:      https://github.com/cucumber/common/tree/suggest/v0.0.1
 [0.0.1]:      https://github.com/cucumber/common/tree/suggest/v0.0.1
 
 <!-- Contributors in alphabetical order -->